pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
dependencies = [
    "gunicorn>=20.1",
    "dj-database-url>=2.0",
    "python-dotenv>=1.0",
    "psycopg2-binary>=2.9",
    "django-bootstrap5>=25.1",
    "whitenoise>=6.9.0",
    "hatchling>=1.27.0",
    "django>=5.2.1",
    "django-filter>=25.1",
    "dotenv>=0.9.9",
    "rollbar>=1.3.0",
    "djangorestframework>=3.16.0",
    "docker>=7.1.0",
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager"]
